Django/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
sqlparse==0.4.3
django-cors-headers==3.13.0
 psycopg2-2.9.3